shift query result downloading into querysupport package
